+++ src/App.jsx
@@ -1,40 +1,14 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import './App.css'
-import { getRandomFact } from './services/facts'
-
-// const CAT_ENDPOINT_IMAGE_URL = `https://cataas.com/cat/says/${firstWord}?size=50&color=red&json=true`
-const CAT_PREFIX_IMAGE_URL = 'https://cataas.com'
-
-function useCatImage ({ fact }) {
-  const [imageUrl, setImageUrl] = useState()
-
-  // Recuperamos la imagen
-  useEffect(() => {
-    if (!fact) return
-    const threeFirstWords = fact.split(' ', 3).join(' ')
-
-    fetch(`https://cataas.com/cat/says/${threeFirstWords}?size=50&color=red&json=true`)
-      .then(res => res.json())
-      .then(response => {
-        const { url } = response
-        setImageUrl(url)
-      })
-  }, [fact])
-
-  return { imageUrl }
-}
+import { useCatFact } from './hooks/useCatFact'
+import { useCatImage } from './hooks/useCatImage'
 
 export function App () {
-  const [fact, setFact] = useState()
+  const { fact, refreshFact } = useCatFact()
   const { imageUrl } = useCatImage({ fact })
 
-  useEffect(() => {
-    getRandomFact().then(newFact => setFact(newFact))
-  }, [])
-
   const handleClick = async () => {
-    const newFact = await getRandomFact()
-    setFact(newFact)
+    refreshFact()
   }
 
   return (
@@ -43,7 +17,7 @@ export function App () {
       <button onClick={handleClick}>Get random Fact</button>
       <section>
         {fact && <p>{fact}</p>}
-        {imageUrl && <img src={`${CAT_PREFIX_IMAGE_URL}${imageUrl}`} alt={`Image extracted using the first word in ${fact}`} />}
+        {imageUrl && <img src={imageUrl} alt={`Image extracted using the first word in ${fact}`} />}
       </section>
     </main>
   )
