+++ src/hooks/useCatImage.js
@@ -1,0 +1,22 @@
+import { useState, useEffect } from 'react'
+
+const CAT_PREFIX_IMAGE_URL = 'https://cataas.com'
+
+export function useCatImage ({ fact }) {
+  const [imageUrl, setImageUrl] = useState()
+
+  // Recuperamos la imagen
+  useEffect(() => {
+    if (!fact) return
+    const threeFirstWords = fact.split(' ', 3).join(' ')
+
+    fetch(`https://cataas.com/cat/says/${threeFirstWords}?size=50&color=red&json=true`)
+      .then(res => res.json())
+      .then(response => {
+        const { url } = response
+        setImageUrl(url)
+      })
+  }, [fact])
+
+  return { imageUrl: `${CAT_PREFIX_IMAGE_URL}${imageUrl}` }
+}
